+++ app/src/main/java/com/example/parcial_2_am_acn4bv_castro_verdugo/ApiObtenerNoticias.java
@@ -62,5 +62,5 @@ public class ApiObtenerNoticias extends AsyncTask<String, Integer, String> {
         super.onPostExecute(s);
         noticiasActivity.cargarDatosDesdeAsyncTask(s);
     }
-}
 
+}
