+++ app/src/main/java/com/example/parcial_2_am_acn4bv_castro_verdugo/InicioSesion.java
@@ -17,24 +17,19 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.firestore.DocumentSnapshot;
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.QueryDocumentSnapshot;
-import com.google.firebase.firestore.QuerySnapshot;
 
 import java.util.List;
 
 public class InicioSesion extends AppCompatActivity {
 
     private FirebaseAuth mAuth;
-    private FirebaseFirestore db;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_inicio_sesion);
         mAuth = FirebaseAuth.getInstance();
-        db = FirebaseFirestore.getInstance();
     }
+
     public void login(String user, String password){
 
         ConnectivityManager connMgr = (ConnectivityManager)
@@ -46,24 +41,8 @@ public class InicioSesion extends AppCompatActivity {
                         @Override
                         public void onComplete(@NonNull Task<AuthResult> task) {
                             if(task.isSuccessful()){
-                                String uid = mAuth.getCurrentUser().getUid();
-                                db.collection("users")
-                                        .whereEqualTo("uidAuth",uid)
-                                        .get()
-                                        .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                                            @Override
-                                            public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                                                if(task.isSuccessful()){
-                                                    for(QueryDocumentSnapshot document: task.getResult()){
-                                                        String nombre = document.getString("nombre");
-                                                        Toast.makeText(getApplicationContext(), "Bienvenido: " + nombre, Toast.LENGTH_SHORT).show();
-                                                        Intent intent = new Intent(getApplicationContext(), Noticias.class);
-                                                        intent.putExtra("nombreUsuario", nombre);
-                                                        startActivity(intent);
-                                                    }
-                                                }
-                                            }
-                                        });
+                                Intent intent = new Intent(getApplicationContext(), Noticias.class);
+                                startActivity(intent);
                             }else{
                                 Toast.makeText(InicioSesion.this, "El inicio de sesión fallo. Usuario y contraseña incorrecto.", Toast.LENGTH_SHORT).show();
                                 Intent intent = new Intent(getApplicationContext(), ErrorInicioSesion.class);
