+++ app/src/main/java/com/example/parcial_2_am_acn4bv_castro_verdugo/MainActivity.java
@@ -5,7 +5,6 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.Toast;
 
@@ -44,21 +43,21 @@ public class MainActivity extends AppCompatActivity {
         super.onStart();
         FirebaseUser currentUser = mAuth.getCurrentUser();
         if(currentUser != null){
-            String uid = currentUser.getUid();
+            String uid = mAuth.getCurrentUser().getUid();
             db.collection("users")
-                    .whereEqualTo("uidAuth",uid)
-
+                    .whereEqualTo("uidAuth", uid)
                     .get()
                     .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
                         @Override
                         public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                            if(task.isSuccessful()){
-                                for(QueryDocumentSnapshot document: task.getResult()){
-                                    String nombre = document.getString("nombre");
-                                    Toast.makeText(getApplicationContext(), "Bienvenido: " + nombre, Toast.LENGTH_SHORT).show();
-                                    Intent intent = new Intent(getApplicationContext(), Noticias.class);
-                                    intent.putExtra("nombreUsuario", nombre);
 
+                            if(task.isSuccessful()){
+                                for(QueryDocumentSnapshot document:task.getResult()){
+                                    String nombre = document.getString("nombre");
+                                    Toast t1 = new Toast(getApplicationContext());
+                                    t1.setText("Bienvenido " + nombre + "!!!!!!");
+                                    t1.show();
+                                    Intent intent = new Intent(getApplicationContext(), Noticias.class);
                                     startActivity(intent);
                                 }
                             }
